Crate storage_service_types

Structs

AccountStatesChunkWithProofRequest

A storage service request for fetching a list of account states at a specified version.

CompleteDataRange

A struct representing a contiguous, non-empty data range (lowest to highest, inclusive) where data is complete (i.e. there are no missing pieces of data).

DataSummary

A summary of the data actually held by the storage service instance.

DegenerateRangeError

EpochEndingLedgerInfoRequest

A storage service request for fetching a list of epoch ending ledger infos.

ProtocolMetadata

A summary of the protocol metadata for the storage service instance, such as the maximum chunk sizes supported for different requests.

ServerProtocolVersion

The protocol version run by this server. Clients request this first to identify what API calls and data requests the server supports.

StorageServerSummary

A storage server summary, containing a summary of the information held by the corresponding server instance. This is useful for identifying the data that a server instance can provide, as well as relevant metadata.

TransactionOutputsWithProofRequest

A storage service request for fetching a transaction output list with a corresponding proof.

TransactionsWithProofRequest

A storage service request for fetching a transaction list with a corresponding proof.

UnexpectedResponseError

Enums

StorageServiceError

A storage service error that can be returned to the client on a failure to process a service request.

StorageServiceMessage

A single storage service message sent or received over DiemNet.

StorageServiceRequest

A storage service request.

StorageServiceResponse

A storage service response.

Type Definitions

Epoch

A type alias for different epochs.

Result